requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
audioread==3.0.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
cloudpickle==1.6.0
cmake==3.26.3
colorama==0.4.4
command-not-found==0.3
cramjam==2.6.2
cryptography==3.4.8
cuda-python==11.7.1
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.30
dbus-python==1.2.18
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dm-tree==0.1.8
docker-pycreds==0.4.0
duplicity==0.8.21
entrypoints==0.4
envs==1.4
executing==0.8.3
Farama-Notifications==0.0.4
fasteners==0.14.1
fastjsonschema==2.15.3
fastparquet==2023.2.0
filelock==3.9.0
fonttools==4.29.1
frozenlist==1.3.3
fs==2.4.12
fsspec==2023.4.0
future==0.18.2
gast==0.5.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
graphviz==0.20.1
grpcio==1.53.0
gym==0.25.1
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.13.1
idna==3.3
importlib-metadata==4.13.0
ipykernel==6.15.1
ipython==8.4.0
ipython_genutils==0.2.0
ipywidgets==7.7.1
jax-jumpy==1.0.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.13
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.1
librosa==0.10.0
lit==16.0.1
llvmlite==0.38.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.5
mutagen==1.46.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nes-py==8.2.1
nest-asyncio==1.5.5
netifaces==0.11.0
networkx==3.1
nnAudio==0.3.2
notebook==6.4.12
numba==0.55.2
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
pandas==2.0.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==3.1.0
ply==3.11
pooch==1.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.17
pycups==2.0.1
pyglet==1.5.0
Pygments==2.12.0
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-slugify==6.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
ray==2.3.1
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
ruptures==1.1.7
scikit-learn==1.1.1
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentry-sdk==1.20.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0
skorch==0.12.1
smmap==5.0.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.4
ssh-import-id==5.11
stack-data==0.3.0
svgwrite==1.4.3
sympy==1.9
systemd-python==234
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.15.0
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.26.1
Tree==0.2.4
triton==2.0.0
typing_extensions==4.3.0
tzdata==2023.3
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.15
usb-creator==0.3.7
virtualenv==20.21.0
wadllib==1.3.6
wandb==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==3.6.1
xdg==5
xkit==0.0.0
yt-dlp==2023.3.4
zipp==1.0.0
